\name{alignmentIndex}
\alias{alignmentIndex}

\title{
  Get the AlignmentIndex from a PacBioCmpH5 object.
}
\description{
  Returns the AlignmentIndex from a PacBioCmpH5 object.
}
\usage{
  alignmentIndex(cmpH5)
}
\arguments{
  \item{cmpH5}{ An object of class PacBioCmpH5}
}
\value{
  Returns a data.frame where each row represents an alignment. The
  specfic columns are documented in the class documentation of
  PacBioCmpH5 \code{\link{PacBioCmpH5-class}}. 
}
\seealso{
  \code{\link{PacBioCmpH5-class}}
}
\examples{
}
\keyword{programming}
