\name{getAlignments}
\alias{getAlignments}
\title{
  Return Alignments
}
\description{
  Return alignments from a class which contains alignments.
}
\usage{
  getAlignments(h5Obj, ...)
}
\arguments{
  \item{h5Obj}{
    An object containing alignments.
  }
  \item{\dots}{
    Extra arguments -- See the particular class documentation for
    specific methods.
  }
}
\details{
  Currently, alignments are two column matrices of mode character. 
}
\value{
  A list of alignment matrices.
}
\examples{
}
\keyword{programming}
