\name{getAttributesForRefSeqs}
\alias{getAttributesForRefSeqs}
\title{
  Get Attributes of Reference Sequences
}
\description{
  Retrieve a list of attributes for the reference sequences represented
  in the cmp.h5 file.
}
\usage{
getAttributesForRefSeqs(cmpH5)
}
\arguments{
  \item{cmpH5}{
    An object of class \code{PacBioCmpH5}
  }
}
\details{
  Each reference sequence represented in the cmp.h5 file.
}
\value{
  A list of lists containing the attributes for each reference sequence.
}
\examples{
}
\keyword{programming}
