\name{getHoleNumbers}
\alias{getHoleNumbers}
\title{
  Get ZMWs Stored in a bas.h5 File.
}
\description{
  Retrieve the ZMW hole numbers stored in the bas.h5 or pls.h5 file.
}
\usage{
  getHoleNumbers(basH5)
}
\arguments{
  \item{basH5}{
    An object of class \code{PacBioCmpH5}
  }
}
\value{
  A vector of integers representing hole numbers.
}
\keyword{programming}
