\name{getInsertionsAndDeletionsByContext}
\alias{getInsertionsAndDeletionsByContext}
\title{
  Return Tables of Insertions and Deletions
}
\description{
  Return tables of insertions and deletion rates. This function might
  have limited appeal, but can be looked at for implementation of more
  complex functionality.
}
\usage{
getInsertionsAndDeletionsByContext(cmpH5, matches = c(AA = 17, CC = 34, GG = 68, TT = 136))
}
\arguments{
  \item{cmpH5}{
  }
  \item{matches}{
  }
}
\details{
}
\value{
  A list of insertions and deletion tables.
}
\examples{
}
\keyword{programming}
