\name{getQualityValue-methods}
\docType{methods}
\alias{getQualityValue-methods}
\title{Methods for Function getQualityValue}
\description{
  Methods for function \code{getQualityValue}, Documentation for quality
  values is stored within the class for which it is being retrieved.
}
\section{Methods}{
\describe{

\item{\code{signature(h5Obj = "PacBioCmpH5")}}{
}

\item{\code{signature(h5Obj = "PacBioPlsH5")}}{
}
}}
\keyword{methods}

