\name{getRefIndexTable}
\alias{getRefIndexTable}
\title{
  Get Reference Index of Offset Table
}
\description{
  Retrieves the reference index table which is used to find the extents
  in the AlignmentIndex correpsonding to each reference. 
}
\usage{
  getRefIndexTable(cmpH5)
}
\arguments{
  \item{cmpH5}{
    An object of class \code{PacBioCmpH5}
  }
}
\details{
  XXX: need to add some details
}
\value{
  A matrix giving the offsets into the alignmentIndex.
}

\examples{
}
\keyword{programming}
