\name{getRegionsTable}
\alias{getRegionsTable}
\alias{getAcqParams}
\alias{getChannelToBaseMap}
\title{
  High-level pulse/base file information.
}
\description{
  Return aspects of the movie acquisition.
}
\usage{
  getRegionsTable(plsH5)
}
\arguments{
  \item{plsH5}{
    Either a PacBioBasH5 or PacBioPlsH5 file.
  }
}
\keyword{classes}
