\name{getQualityValue}
\alias{getQualityValue}
\title{
  Return Classifier Quality Values
}
\description{
  Return classifier quality values from a class/file which contains
  them. 
}
\usage{
  getQualityValue(h5Obj, ...)
}
\arguments{
  \item{h5Obj}{
    An object of class \code{\link{PacBioCmpH5-class}} or
    \code{\link{PacBioBasH5-class}} 
  }
  \item{\dots}{
    Extra arguments -- See class documentation for \code{h5Obj}
  }
}
\value{
  A list of vectors of quality values.
}
\examples{
}
\keyword{programming}
